requirements.txt
torch
torchvision
accelerate
datasets
peft
sentencepiece
pillow
streamlit
pyyaml
pandas
matplotlib
jupyter
ipykernel
qwen-vl-utils==0.0.8
git+https://github.com/huggingface/transformers